64bit: fix warnings
* device/net_io.h (net_set_filter, ethernet_priority): Add prototypes.
* device/subrs.h: Include <device/if_hdr.h>.
(if_init_queues): Add prototype.
* i386/i386/model_dep.h (machine_relax): Add prototype.
* i386/i386/trap.c (i386_astintr): Move mycpu variable definition to
where it is used.
* i386/i386at/model_dep.c (i386at_init): Likewise for nb_direct, addr,
delta.
* i386/xen/xen.c (return_to_iret): Change type to char[].
* xen/console.c: Include <i386at/kd.h>.
* xen/evt.c (hyp_evt_handler): Cast NEVNT to int.
* xen/grant.c: Include <model_dep.h>.
(hyp_grant_takeback, hyp_grant_init): Fix print format.
* xen/net.c: Include <device/subrs.h>.
(paranoia): Remove variable.
(hyp_net_init, device_close, device_open): Cast nd - vif_data to int.
Fix print format.
* xen/store.c (store_put): Cast sizeof to int.
* xen/time.c: Include "xen.h".
* xen/xen.h (hypclock_machine_intr): Add prototype.

+++ xen/evt.c
@@ -106,7 +106,7 @@ void hyp_intrinit() {
 
 void hyp_evt_handler(evtchn_port_t port, void (*handler)(), int unit, spl_t spl) {
 	if (port > NEVNT)
-		panic("event channel port %d > %d not supported\n", port, NEVNT);
+		panic("event channel port %d > %d not supported\n", port, (int) NEVNT);
 	intpri[port] = spl;
 	iunit[port] = unit;
 	form_int_mask();
